Refactor business create models into dedicated files
