+++ src/main/kotlin/no/elhub/auth/features/documents/common/DocumentBusinessHandler.kt
@@ -4,10 +4,8 @@ import arrow.core.Either
 import no.elhub.auth.features.businessprocesses.BusinessProcessError
 import no.elhub.auth.features.businessprocesses.changeofbalancesupplier.ChangeOfBalanceSupplierBusinessHandler
 import no.elhub.auth.features.businessprocesses.moveinandchangeofbalancesupplier.MoveInAndChangeOfBalanceSupplierBusinessHandler
-import no.elhub.auth.features.common.party.AuthorizationParty
 import no.elhub.auth.features.documents.AuthorizationDocument
 import no.elhub.auth.features.documents.create.command.DocumentCommand
-import no.elhub.auth.features.documents.create.dto.SupportedLanguageDTO
 import no.elhub.auth.features.grants.common.CreateGrantProperties
 
 class ProxyDocumentBusinessHandler(
@@ -35,22 +33,3 @@ interface DocumentBusinessHandler {
 
     fun getCreateGrantProperties(document: AuthorizationDocument): CreateGrantProperties
 }
-
-data class CreateDocumentBusinessModel(
-    val authorizedParty: AuthorizationParty,
-    val documentType: AuthorizationDocument.Type,
-    val requestedBy: AuthorizationParty,
-    val requestedFrom: AuthorizationParty,
-    val requestedTo: AuthorizationParty,
-    val meta: CreateDocumentBusinessMeta,
-)
-
-data class CreateDocumentBusinessMeta(
-    val requestedFromName: String,
-    val requestedForMeteringPointId: String,
-    val requestedForMeteringPointAddress: String,
-    val balanceSupplierName: String,
-    val balanceSupplierContractName: String,
-    val moveInDate: kotlinx.datetime.LocalDate? = null,
-    val language: SupportedLanguageDTO = SupportedLanguageDTO.DEFAULT,
-)
