+++ src/main/kotlin/no/elhub/auth/features/requests/common/ProxyRequestBusinessHandler.kt
@@ -1,11 +1,9 @@
 package no.elhub.auth.features.requests.common
 
 import arrow.core.Either
-import kotlinx.datetime.LocalDate
 import no.elhub.auth.features.businessprocesses.BusinessProcessError
 import no.elhub.auth.features.businessprocesses.changeofbalancesupplier.ChangeOfBalanceSupplierBusinessHandler
 import no.elhub.auth.features.businessprocesses.moveinandchangeofbalancesupplier.MoveInAndChangeOfBalanceSupplierBusinessHandler
-import no.elhub.auth.features.common.party.AuthorizationParty
 import no.elhub.auth.features.grants.common.CreateGrantProperties
 import no.elhub.auth.features.requests.AuthorizationRequest
 import no.elhub.auth.features.requests.create.command.RequestCommand
@@ -34,22 +32,3 @@ interface RequestBusinessHandler {
     suspend fun validateAndReturnRequestCommand(createRequestModel: CreateRequestBusinessModel): Either<BusinessProcessError, RequestCommand>
     fun getCreateGrantProperties(request: AuthorizationRequest): CreateGrantProperties
 }
-
-data class CreateRequestBusinessModel(
-    val authorizedParty: AuthorizationParty,
-    val requestType: AuthorizationRequest.Type,
-    val requestedBy: AuthorizationParty,
-    val requestedFrom: AuthorizationParty,
-    val requestedTo: AuthorizationParty,
-    val meta: CreateRequestBusinessMeta,
-)
-
-data class CreateRequestBusinessMeta(
-    val requestedFromName: String,
-    val requestedForMeteringPointId: String,
-    val requestedForMeteringPointAddress: String,
-    val balanceSupplierName: String,
-    val balanceSupplierContractName: String,
-    val moveInDate: LocalDate? = null,
-    val redirectURI: String? = null,
-)
